Applicaiton written using LAMP and Bootstrap
Collegium (unfnished) - mobile friendly site for college students

DECODER - same game as the Android version but for the web

Secure Web Registeration and Login -
Using LAMP stack, I created a registeration/login system that offers
captcha, email verification,  temp database for temp information as well
as salting confidential information in main database. It is also secure
from SQL injection and a few other hacks.

There is also a Bashscript designed to be ran in case of a disaster to
recover the entire system including the LAMP stack and database and the
web pages all automatically in linux environment.